Typing sortByOrder function
